- Generic classes now each have their own files. - Fixed a refactoring error that unintentionally renamed MenuItem to MenuItemProvider. - Fixed .editorconfig indent size. - Reformatted code to ensure it complies with the ReSharper and editor config settings.
