+++ Examples/Jcd.RichEnumerations.Examples.CardSelector/Jcd.RichEnumerations.Examples.CardSelector.csproj
@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
+   <PropertyGroup>
+      <OutputType>Exe</OutputType>
+      <TargetFramework>net8.0</TargetFramework>
+      <ImplicitUsings>enable</ImplicitUsings>
+      <Nullable>enable</Nullable>
+   </PropertyGroup>
 
-    <ItemGroup>
-        <ProjectReference Include="..\..\Jcd.RichEnumerations\Jcd.RichEnumerations.csproj"/>
-    </ItemGroup>
+   <ItemGroup>
+      <ProjectReference Include="..\..\Jcd.RichEnumerations\Jcd.RichEnumerations.csproj"/>
+   </ItemGroup>
 
 </Project>
