+++ Jcd.RichEnumerations/DescribedRichEnum{TValue, TEnum}.cs
@@ -1,22 +1,9 @@
 ﻿using System;
 
+// ReSharper disable UnusedType.Global
 // ReSharper disable HeapView.PossibleBoxingAllocation
 // ReSharper disable HeapView.ObjectAllocation
-// ReSharper disable MemberCanBePrivate.Global
-
 namespace Jcd.RichEnumerations;
-
-#pragma warning disable SA1402 // FileMayOnlyContainASingleType
-
-/// <summary>
-/// A <see cref="RichEnum{TEnum}"/> with a Description property.
-/// </summary>
-/// <param name="value">the value for this instance. These must be unique.</param>
-/// <param name="description">The description for the instance. These should be unique.</param>
-/// <typeparam name="TEnum">The type deriving from DEnumeration</typeparam>
-public abstract class DescribedRichEnum<TEnum>(int value, string description)
-   : DescribedRichEnum<int, TEnum>(value, description)
-   where TEnum : DescribedRichEnum<TEnum>, IComparable<TEnum>;
 
 /// <summary>
 /// A <see cref="RichEnum{TValue,TEnum}"/> with a Description property.
