+++ Jcd.RichEnumerations/SortableRichEnum{TValue, TEnum}.cs
@@ -1,19 +1,9 @@
 ﻿using System;
 
+// ReSharper disable UnusedType.Global
 // ReSharper disable HeapView.PossibleBoxingAllocation
 // ReSharper disable HeapView.ObjectAllocation
-
 namespace Jcd.RichEnumerations;
-
-#pragma warning disable SA1402 // FileMayOnlyContainASingleType
-
-/// <summary>
-/// A rich enum that can be sorted with a call to .Sort.
-/// </summary>
-/// <param name="value">The value of the instance. This must be unique.</param>
-/// <typeparam name="TEnum">The type deriving from SortableEnum{TEnum}</typeparam>
-public abstract class SortableRichEnum<TEnum>(int value) : SortableRichEnum<int, TEnum>(value)
-   where TEnum : SortableRichEnum<TEnum>;
 
 /// <summary>
 /// A rich enum that can be sorted with a call to .Sort.
@@ -51,7 +41,8 @@ public abstract class SortableRichEnum<TValue, TEnum>(TValue value) : SortableRi
    }
 
    /// <summary>
-   /// Compares the Value of the current <see cref="SortableRichEnum{TValue,TEnum}"/> instance to the target instance for equality.
+   /// Compares the Value of the current <see cref="SortableRichEnum{TValue,TEnum}"/> instance to the
+   /// target instance for equality.
    /// </summary>
    /// <param name="other">The other instance to compare to</param>
    /// <returns>True if equivalent.</returns>
@@ -160,10 +151,14 @@ public abstract class SortableRichEnum<TValue, TEnum>(TValue value) : SortableRi
    public static explicit operator TValue(SortableRichEnum<TValue, TEnum> e) { return e.Value; }
 
    /// <summary>
-   /// Explicit conversion from the underlying data type to an <see cref="SortableRichEnum{TValue,TEnum}"/> instance. This performs a lookup on the known values.
+   /// Explicit conversion from the underlying data type to an <see cref="SortableRichEnum{TValue,TEnum}"/>
+   /// instance. This performs a lookup on the known values.
    /// </summary>
    /// <param name="value">The value to convert.</param>
-   /// <returns>The instance of the <see cref="SortableRichEnum{TValue,TEnum}"/>, if it exists. Throws an exception, otherwise.</returns>
+   /// <returns>
+   /// The instance of the <see cref="SortableRichEnum{TValue,TEnum}"/>, if it exists.
+   /// Throws an exception, otherwise.
+   /// </returns>
    /// <exception cref="ArgumentException">If a value not present in the enum is provided.</exception>
    public static explicit operator SortableRichEnum<TValue, TEnum>(TValue value)
    {
